requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
Authlib==1.6.3
Automat==25.4.16
babel==2.17.0
backoff==2.2.1
bce-python-sdk==0.9.45
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cohere==5.17.0
colorama==0.4.6
comm==0.2.3
constantly==23.10.4
cos_python_sdk_v5==1.9.38
crcmod==1.7
cryptography==45.0.5
cssselect==1.3.0
dataclasses-json==0.6.7
ddgr==2.2
ddgs==9.5.5
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
Django==5.2.5
django-cors-headers==4.7.0
django-redis==6.0.0
duckduckgo_search==8.1.1
emoji==2.14.1
et_xmlfile==2.0.0
executing==2.2.1
faiss-cpu==1.12.0
fastavro==1.12.0
fastjsonschema==2.21.2
filelock==3.18.0
filetype==1.2.0
Flask==3.1.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-tools==1.74.0
h11==0.16.0
hf-xet==1.1.9
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.34.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itemadapter==0.12.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jwt==1.4.0
langchain==0.3.27
langchain-cohere==0.4.5
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-huggingface==0.3.1
langchain-openai==0.3.32
langchain-pinecone==0.2.12
langchain-text-splitters==0.3.11
langchain-weaviate==0.0.5
langdetect==1.0.9
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.20
lark==1.2.2
longchain==0.4.2
lxml==6.0.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.6.4
multiprocess==0.70.18
mypy_extensions==1.1.0
mysql==0.0.3
mysqlclient==2.2.7
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
olefile==0.47
openai==1.102.0
openpyxl==3.1.5
orjson==3.11.3
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.2
pandocfilters==1.5.1
parsel==1.10.0
parso==0.8.5
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
platformdirs==4.4.0
primp==0.15.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
Protego==0.5.0
protobuf==6.32.0
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyDispatcher==2.0.7
Pygments==2.19.2
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-json-logger==3.3.0
python-magic==0.4.27
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
pywin32==311
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.0.2
qianfan==0.4.12.3
queuelib==1.8.0
rank-bm25==0.2.2
RapidFuzz==3.14.0
redis==6.4.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
Scrapy==2.13.3
Send2Trash==1.8.3
sentence-transformers==5.1.0
service-identity==24.2.0
setuptools==80.9.0
shellingham==1.5.4
simsimd==6.5.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.14.0
tcvdb_text==1.1.2
tcvectordb==1.8.4
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.11.0
tinycss2==1.4.0
tldextract==5.3.0
tokenizers==0.22.0
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.0
Twisted==25.5.0
typer==0.17.3
types-python-dateutil==2.9.0.20250822
types-PyYAML==6.0.12.20250822
types-requests==2.32.4.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
unstructured==0.18.14
unstructured-client==0.42.3
uri-template==1.3.0
urllib3==2.5.0
validators==0.35.0
w3lib==2.3.1
wcwidth==0.2.13
weaviate-client==4.16.9
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.3
xlsxwriter==3.2.5
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.20.1
zope.interface==7.2
zstandard==0.24.0
